@@ -20,14 +20,12 @@ classifiers = [
     'License :: OSI Approved :: MIT License',
     'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
     'Programming Language :: Python',
-    'Programming Language :: Python :: 3.8',
-    'Programming Language :: Python :: 3.9',
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
     'Programming Language :: Python :: 3.12',
 ]
 keywords = ['workflow', 'multithreaded', 'rabbitmq']
-requires-python = '>=3.8'
+requires-python = '>=3.10'
 dependencies = [
     'kiwipy[rmq]~=0.8.5',
     'nest_asyncio~=1.5,>=1.5.1',
